fix: Extract renamed repo name for use as team name
Extract the renamed repo name from HTML_URL to use as the team name instead of the API repo name
This allows the tier name (e.g. Bears, Geckos, Toucans) to be extracted from the API repo name
but the team name will be from the HTML_URL which is the name commonly used by the team.

Resolves: N/a
See also: N/a